+++ zfit/binned.py
@@ -3,4 +3,4 @@ from __future__ import annotations
 
 from ._variables.axis import Binnings, RegularBinning, VariableBinning
 
-__all__ = ["RegularBinning", "VariableBinning", "Binnings"]
+__all__ = ["Binnings", "RegularBinning", "VariableBinning"]