+++ zfit/loss.py
@@ -10,13 +10,13 @@ from ._loss.binnedloss import (
 from .core.loss import BaseLoss, ExtendedUnbinnedNLL, SimpleLoss, UnbinnedNLL
 
 __all__ = [
-    "ExtendedUnbinnedNLL",
-    "UnbinnedNLL",
-    "BinnedNLL",
-    "ExtendedBinnedNLL",
     "BaseLoss",
-    "SimpleLoss",
-    "ExtendedBinnedNLL",
     "BinnedChi2",
+    "BinnedNLL",
     "ExtendedBinnedChi2",
+    "ExtendedBinnedNLL",
+    "ExtendedBinnedNLL",
+    "ExtendedUnbinnedNLL",
+    "SimpleLoss",
+    "UnbinnedNLL",
 ]