+++ zfit/plot.py
@@ -4,4 +4,4 @@ from __future__ import annotations
 
 from .util.plotter import PDFPlotter, SumCompPlotter, ZfitPDFPlotter, plot_model_pdfV1, plot_sumpdf_components_pdfV1
 
-__all__ = ["plot_model_pdfV1", "plot_sumpdf_components_pdfV1", "ZfitPDFPlotter", "PDFPlotter", "SumCompPlotter"]
+__all__ = ["PDFPlotter", "SumCompPlotter", "ZfitPDFPlotter", "plot_model_pdfV1", "plot_sumpdf_components_pdfV1"]